Work around authentiication issue.
Currently it is not possible to access a blob purely by url. For that
some form of authentication is needed. A connection string is the
easiest. The blob client class hoewever can not take both
connectionstring and bloburl. As a workaround I parse container and blob
name from the blobUrl and inject them into a client configured with a
connection string.

Later we can create a shared access key and pass that in, instead of
this workaround.